fix bug where the zoom of quadkey doesn't always  work
